+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * Add constraint on password confirmation
 * Add rating form
 ### Fix
+* UI Homepage
+* Added signup page
+* Added modal to select campus to create a new ride from homepage
 * Fix infosec issue with X-Frame header
 ## Version 1.4.3
 ### Features